refactor: sns 목록 공통화
