+++ src/components/Navigation/MobileNavigation.tsx
@@ -1,32 +1,8 @@
 import Link from "next/link";
 import { useEffect, useState } from "react";
-import {
-  DddIcon,
-  MenuIcon,
-  CloseIcon,
-  InstagramIcon,
-  LinkedInIcon,
-  GithubIcon,
-} from "@/components/svgs";
+import { DddIcon, MenuIcon, CloseIcon } from "@/components/svgs";
+import snsList from "@/components/common/snsList";
 import navigationList from "@/fixtures/navigationList.json";
-
-const snsList = [
-  {
-    name: "Instagram",
-    icon: <InstagramIcon />,
-    link: "https://www.instagram.com/dynamic_ddd?utm_source=ig_web_button_share_sheet&igsh=ZDNlZDc0MzIxNw==",
-  },
-  {
-    name: "LinkedIn",
-    icon: <LinkedInIcon />,
-    link: "https://www.linkedin.com/company/dddcommunity",
-  },
-  {
-    name: "Github",
-    icon: <GithubIcon />,
-    link: "https://github.com/DDD-Community",
-  },
-];
 
 export default function MobileNavigation() {
   const [isMenuOpened, setMenuOpened] = useState(false);
