+++ src/components/common/snsList.tsx
@@ -1,0 +1,21 @@
+import { InstagramIcon, LinkedInIcon, GithubIcon } from "@/components/svgs";
+
+const snsList = [
+  {
+    name: "Instagram",
+    icon: <InstagramIcon />,
+    link: "https://www.instagram.com/dynamic_ddd?utm_source=ig_web_button_share_sheet&igsh=ZDNlZDc0MzIxNw==",
+  },
+  {
+    name: "LinkedIn",
+    icon: <LinkedInIcon />,
+    link: "https://www.linkedin.com/company/dddcommunity",
+  },
+  {
+    name: "Github",
+    icon: <GithubIcon />,
+    link: "https://github.com/DDD-Community",
+  },
+];
+
+export default snsList;
